+++ geotrek/common/parsers.py
@@ -6,7 +6,6 @@ import requests
 from requests.auth import HTTPBasicAuth
 import xlrd
 import xml.etree.ElementTree as ET
-import json
 import urllib2
 
 from ftplib import FTP
@@ -27,6 +26,7 @@ from modeltranslation.fields import TranslationField
 from modeltranslation.translator import translator, NotRegistered
 from paperclip.models import Attachment, attachment_upload
 
+from geotrek.authent.models import default_structure
 from geotrek.common.models import FileType
 
 
@@ -107,6 +107,22 @@ class Parser(object):
         warnings = self.warnings.setdefault(key, [])
         warnings.append(msg)
 
+    def get_part(self, dst, src, val):
+        if not src:
+            return val
+        if val is None:
+            return None
+        if '.' in src:
+            part, left = src.split('.', 1)
+        else:
+            part, left = src, ''
+        if part.isdigit():
+            return self.get_part(dst, left, val[int(part)])
+        elif part == '*':
+            return [self.get_part(dst, left, subval) for subval in val]
+        else:
+            return self.get_part(dst, left, val[part])
+
     def get_val(self, row, dst, src):
         if hasattr(src, '__iter__'):
             val = []
@@ -117,18 +133,13 @@ class Parser(object):
                     if self.warn_on_missing_fields:
                         self.add_warning(unicode(warning))
                     val.append(None)
+            return val
         else:
-            val = row
-            for part in src.split('.'):
-                try:
-                    if part.isdigit():
-                        val = val[int(part)]
-                    else:
-                        val = val[part]
-                except (KeyError, IndexError):
-                    required = u"required " if self.field_options.get(dst, {}).get('required', False) else ""
-                    raise ValueImportError(_(u"Missing {required}field '{src}'").format(required=required, src=src))
-        return val
+            try:
+                return self.get_part(dst, src, row)
+            except (KeyError, IndexError):
+                required = u"required " if self.field_options.get(dst, {}).get('required', False) else ""
+                raise ValueImportError(_(u"Missing {required}field '{src}'").format(required=required, src=src))
 
     def apply_filter(self, dst, src, val):
         field = self.model._meta.get_field_by_name(dst)[0]
@@ -160,20 +171,12 @@ class Parser(object):
             raise RowImportError(_(u"Blank value not allowed for field '{src}'".format(src=src)))
         setattr(self.obj, dst, val)
 
-    def parse_field(self, dst, src, val):
+    def parse_real_field(self, dst, src, val):
         """Returns True if modified"""
         if hasattr(self, 'filter_{0}'.format(dst)):
-            try:
-                val = getattr(self, 'filter_{0}'.format(dst))(src, val)
-            except ValueImportError as warning:
-                self.add_warning(unicode(warning))
-                return False
+            val = getattr(self, 'filter_{0}'.format(dst))(src, val)
         else:
-            try:
-                val = self.apply_filter(dst, src, val)
-            except ValueImportError as warning:
-                self.add_warning(unicode(warning))
-                return False
+            val = self.apply_filter(dst, src, val)
         if hasattr(self.obj, dst):
             if dst in self.m2m_fields or dst in self.m2m_constant_fields:
                 old = set(getattr(self.obj, dst).all())
@@ -192,30 +195,35 @@ class Parser(object):
             self.set_value(dst, src, val)
             return True
 
+    def parse_field(self, row, dst, src, updated, non_field):
+        if dst in self.constant_fields:
+            val = self.constant_fields[dst]
+        elif dst in self.m2m_constant_fields:
+            val = self.m2m_constant_fields[dst]
+        else:
+            src = self.normalize_src(src)
+            val = self.get_val(row, dst, src)
+        if non_field:
+            modified = self.parse_non_field(dst, src, val)
+        else:
+            modified = self.parse_real_field(dst, src, val)
+        if modified:
+            updated.append(dst)
+            if dst in self.translated_fields:
+                lang = translation.get_language()
+                updated.append('{field}_{lang}'.format(field=dst, lang=lang))
+
     def parse_fields(self, row, fields, non_field=False):
         updated = []
         for dst, src in fields.items():
-            if dst in self.constant_fields or dst in self.m2m_constant_fields:
-                val = src
-            else:
-                src = self.normalize_src(src)
-                try:
-                    val = self.get_val(row, dst, src)
-                except ValueImportError as warning:
-                    if self.field_options.get(dst, {}).get('required', False):
-                        raise RowImportError(warning)
-                    if self.warn_on_missing_fields:
-                        self.add_warning(unicode(warning))
-                    continue
-            if non_field:
-                modified = self.parse_non_field(dst, src, val)
-            else:
-                modified = self.parse_field(dst, src, val)
-            if modified:
-                updated.append(dst)
-                if dst in self.translated_fields:
-                    lang = translation.get_language()
-                    updated.append('{field}_{lang}'.format(field=dst, lang=lang))
+            try:
+                self.parse_field(row, dst, src, updated, non_field)
+            except ValueImportError as warning:
+                if self.field_options.get(dst, {}).get('required', False):
+                    raise RowImportError(warning)
+                if self.warn_on_missing_fields:
+                    self.add_warning(unicode(warning))
+                continue
         return updated
 
     def parse_obj(self, row, operation):
@@ -270,13 +278,13 @@ class Parser(object):
             objects = self.model.objects.filter(**eid_kwargs)
         if len(objects) == 0 and self.update_only:
             if self.warn_on_missing_objects:
-                self.add_warning(_(u"Bad value '{eid_val}' for field '{eid_src}'. No trek with this identifier").format(eid_val=self.eid_val, eid_src=self.eid_src))
+                self.add_warning(_(u"Bad value '{eid_val}' for field '{eid_src}'. No object with this identifier").format(eid_val=self.eid_val, eid_src=self.eid_src))
             return
         elif len(objects) == 0:
             objects = [self.model(**eid_kwargs)]
             operation = u"created"
         elif len(objects) >= 2 and not self.duplicate_eid_allowed:
-            self.add_warning(_(u"Bad value '{eid_val}' for field '{eid_src}'. Multiple treks with this identifier").format(eid_val=self.eid_val, eid_src=self.eid_src))
+            self.add_warning(_(u"Bad value '{eid_val}' for field '{eid_src}'. Multiple objects with this identifier").format(eid_val=self.eid_val, eid_src=self.eid_src))
             return
         else:
             operation = u"updated"
@@ -293,7 +301,7 @@ class Parser(object):
             'nb_lines': self.line,
             'nb_created': self.nb_created,
             'nb_updated': self.nb_updated,
-            'nb_deleted': len(self.to_delete) if self.delete else 0,
+            'nb_deleted': len(self.to_delete) if self.delete else None,
             'nb_unmodified': self.nb_unmodified,
             'warnings': self.warnings,
         }
@@ -336,7 +344,8 @@ class Parser(object):
     def filter_m2m(self, src, val, model, field, mapping=None, partial=False, create=False):
         if not val:
             return []
-        val = val.split(self.separator)
+        if self.separator:
+            val = val.split(self.separator)
         dst = []
         for subval in val:
             subval = subval.strip()
@@ -357,7 +366,27 @@ class Parser(object):
         return dst
 
     def start(self):
-        self.to_delete = set(self.model.objects.values_list('pk', flat=True))
+        # FIXME: use mapping if it exists
+        kwargs = {}
+        for dst, val in self.constant_fields.iteritems():
+            field = self.model._meta.get_field_by_name(dst)[0]
+            if isinstance(field, models.ForeignKey):
+                natural_key = self.natural_keys[dst]
+                try:
+                    kwargs[dst] = field.rel.to.objects.get(**{natural_key: val})
+                except field.rel.to.DoesNotExist:
+                    raise GlobalImportError(_(u"{model} '{val}' does not exists in Geotrek-Admin. Please add it").format(model=field.rel.to._meta.verbose_name.title(), val=val))
+            else:
+                kwargs[dst] = val
+        for dst, val in self.m2m_constant_fields.iteritems():
+            assert not self.separator or self.separator not in val
+            field = self.model._meta.get_field_by_name(dst)[0]
+            natural_key = self.natural_keys[dst]
+            try:
+                kwargs[dst] = field.rel.to.objects.get(**{natural_key: subval for subval in val})
+            except field.rel.to.DoesNotExist:
+                raise GlobalImportError(_(u"{model} '{val}' does not exists in Geotrek-Admin. Please add it").format(model=field.rel.to._meta.verbose_name.title(), val=val))
+        self.to_delete = set(self.model.objects.filter(**kwargs).values_list('pk', flat=True))
 
     def end(self):
         if self.delete:
@@ -377,9 +406,9 @@ class Parser(object):
             try:
                 self.parse_row(row)
             except Exception as e:
-                self.add_warning(unicode(e))
                 if settings.DEBUG:
                     raise
+                self.add_warning(unicode(e))
         self.end()
 
 
@@ -455,23 +484,15 @@ class AttachmentParserMixin(object):
     def start(self):
         super(AttachmentParserMixin, self).start()
         try:
-            self.filetype = FileType.objects.get(type=self.filetype_name)
+            self.filetype = FileType.objects.get(type=self.filetype_name, structure=default_structure())
         except FileType.DoesNotExist:
             raise GlobalImportError(_(u"FileType '{name}' does not exists in Geotrek-Admin. Please add it").format(name=self.filetype_name))
         self.creator, created = get_user_model().objects.get_or_create(username='import', defaults={'is_active': False})
-        self.attachments_to_delete = {obj.pk: set(Attachment.objects.attachments_for_object(obj)) for obj in self.model.objects.all()}
-
-    def end(self):
-        if self.delete_attachments:
-            for atts in self.attachments_to_delete.itervalues():
-                for att in atts:
-                    att.delete()
-        super(AttachmentParserMixin, self).end()
 
     def filter_attachments(self, src, val):
         if not val:
             return []
-        return [(subval.strip(), '', '') for subval in val.split(self.separator)]
+        return [(subval.strip(), '', '') for subval in val.split(self.separator) if subval.strip()]
 
     def has_size_changed(self, url, attachment):
         try:
@@ -493,20 +514,39 @@ class AttachmentParserMixin(object):
             return False
         return True
 
+    def download_attachment(self, url):
+        if url[:6] == 'ftp://':
+            try:
+                response = urllib2.urlopen(url)
+            except:
+                self.add_warning(_(u"Failed to download '{url}'").format(url=url))
+                return None
+            return response.read()
+        else:
+            try:
+                response = requests.get(url)
+            except requests.exceptions.RequestException as e:
+                raise ValueImportError('Failed to load attachment: ' + unicode(e))
+            if response.status_code != requests.codes.ok:
+                self.add_warning(_(u"Failed to download '{url}'").format(url=url))
+                return None
+            return response.content
+
     def save_attachments(self, src, val):
         updated = False
+        attachments_to_delete = list(Attachment.objects.attachments_for_object(self.obj))
         for url, legend, author in self.filter_attachments(src, val):
             url = self.base_url + url
             legend = legend or u""
             author = author or u""
             name = os.path.basename(url)
             found = False
-            for attachment in self.attachments_to_delete.get(self.obj.pk, set()):
+            for attachment in attachments_to_delete:
                 upload_name, ext = os.path.splitext(attachment_upload(attachment, name))
                 existing_name = attachment.attachment_file.name
                 if re.search(ur"^{name}(_\d+)?{ext}$".format(name=upload_name, ext=ext), existing_name) and not self.has_size_changed(url, attachment):
                     found = True
-                    self.attachments_to_delete[self.obj.pk].remove(attachment)
+                    attachments_to_delete.remove(attachment)
                     if author != attachment.author or legend != attachment.legend:
                         attachment.author = author
                         attachment.legend = legend
@@ -515,19 +555,10 @@ class AttachmentParserMixin(object):
                     break
             if found:
                 continue
-            if url[:6] == 'ftp://':
-                try:
-                    response = urllib2.urlopen(url)
-                except:
-                    self.add_warning(_(u"Failed to download '{url}'").format(url=url))
-                    continue
-                f = ContentFile(response.read())
-            else:
-                response = requests.get(url)
-                if response.status_code != requests.codes.ok:
-                    self.add_warning(_(u"Failed to download '{url}'").format(url=url))
-                    continue
-                f = ContentFile(response.content)
+            content = self.download_attachment(url)
+            if content is None:
+                continue
+            f = ContentFile(content)
             attachment = Attachment()
             attachment.content_object = self.obj
             attachment.attachment_file.save(name, f, save=False)
@@ -537,6 +568,9 @@ class AttachmentParserMixin(object):
             attachment.legend = legend
             attachment.save()
             updated = True
+        if self.delete_attachments:
+            for att in attachments_to_delete:
+                att.delete()
         return updated
 
 
@@ -568,7 +602,7 @@ class TourInSoftParser(AttachmentParserMixin, Parser):
     def filter_attachments(self, src, val):
         if not val:
             return []
-        return [subval.split('||') for subval in val.split('##') if subval.split('||') != ['', '', '']]
+        return [subval.split('||') for subval in val.split('##') if subval.split('||')[0]]
 
 
 class TourismSystemParser(AttachmentParserMixin, Parser):
@@ -603,49 +637,6 @@ class TourismSystemParser(AttachmentParserMixin, Parser):
             except KeyError:
                 name = None
             result.append((subval['URL'], name, None))
-        return result
-
-    def normalize_field_name(self, name):
-        return name
-
-
-class SitraParser(AttachmentParserMixin, Parser):
-    url = 'http://api.sitra-tourisme.com/api/v002/recherche/list-objets-touristiques/'
-
-    @property
-    def items(self):
-        return self.root['objetsTouristiques']
-
-    def next_row(self):
-        size = 100
-        skip = 0
-        while True:
-            params = {
-                'apiKey': self.api_key,
-                'projetId': self.project_id,
-                'selectionIds': [self.selection_id],
-                'count': size,
-                'first': skip,
-            }
-            response = requests.get(self.url, params={'query': json.dumps(params)})
-            if response.status_code != 200:
-                raise GlobalImportError(_(u"Failed to download {url}. HTTP status code {status_code}").format(url=self.url, status_code=response.status_code))
-            self.root = response.json()
-            self.nb = int(self.root['numFound'])
-            for row in self.items:
-                yield row
-            skip += size
-            if skip >= self.nb:
-                return
-
-    def filter_attachments(self, src, val):
-        result = []
-        for subval in val or []:
-            if 'nom' in subval:
-                name = subval['nom']['libelleFr']
-            else:
-                name = None
-            result.append((subval['traductionFichiers'][0]['url'], name, None))
         return result
 
     def normalize_field_name(self, name):