+++ geotrek/settings/base.py
@@ -380,7 +380,7 @@ PAPERCLIP_CONFIG = {
 
 
 # Data projection
-SRID = 3857
+SRID = 2154  # Lambert-93 for Metropolitan France
 
 # API projection (client-side), can differ from SRID (database). Leaflet requires 4326.
 API_SRID = 4326
@@ -425,10 +425,10 @@ ALTIMETRIC_AREA_MARGIN = 0.15
 
 # Let this be defined at instance-level
 LEAFLET_CONFIG = {
-    'SRID': SRID,
+    'SRID': 3857,
     'TILES': [
-        ('Scan', 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',),
-        ('Ortho', 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.jpg'),
+        ('Scan', 'http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', '(c) OpenStreetMap Contributors'),
+        ('Ortho', 'http://oatile1.mqcdn.com/tiles/1.0.0/sat/{z}/{x}/{y}.png', '(c) MapQuest'),
     ],
     'TILES_EXTENT': SPATIAL_EXTENT,
     # Extent in API projection (Leaflet view default extent)
@@ -524,7 +524,6 @@ TREK_POINTS_OF_REFERENCE_ENABLED = True
 TREK_EXPORT_POI_LIST_LIMIT = 14
 TREK_EXPORT_INFORMATION_DESK_LIST_LIMIT = 2
 
-TREK_DAY_DURATION = 10  # Max duration to be done in one day
 TREK_ICON_SIZE_POI = 18
 TREK_ICON_SIZE_SERVICE = 18
 TREK_ICON_SIZE_PARKING = 18