+++ geotrek/tourism/management/commands/sync_rando.py
@@ -39,20 +39,24 @@ class Command(BaseCommand):
         super(Command, self).handle(*args, **options)
 
     def sync_content(self, lang, content):
-        self.sync_pdf(lang, content)
+        if not self.skip_pdf:
+            view = tourism_views.TouristicContentDocumentPublic.as_view(model=type(content))
+            self.sync_object_view(lang, content, view, '{obj.slug}.pdf')
 
         for picture, resized in content.resized_pictures:
             self.sync_media_file(lang, resized)
 
     def sync_event(self, lang, event):
-        self.sync_pdf(lang, event)
+        if not self.skip_pdf:
+            view = tourism_views.TouristicEventDocumentPublic.as_view(model=type(event))
+            self.sync_object_view(lang, event, view, '{obj.slug}.pdf')
 
         for picture, resized in event.resized_pictures:
             self.sync_media_file(lang, resized)
 
     def sync_tourism(self, lang):
-        self.sync_geojson(lang, tourism_views.TouristicContentViewSet, 'touristiccontents')
-        self.sync_geojson(lang, tourism_views.TouristicEventViewSet, 'touristicevents',
+        self.sync_geojson(lang, tourism_views.TouristicContentViewSet, 'touristiccontents.geojson')
+        self.sync_geojson(lang, tourism_views.TouristicEventViewSet, 'touristicevents.geojson',
                           params={'ends_after': timezone.now().strftime('%Y-%m-%d')})
 
         # reopen global zip (closed after trekking sync)
@@ -98,6 +102,14 @@ class Command(BaseCommand):
             events = events.filter(source__name__in=self.source)
         for event in events:
             self.sync_event(lang, event)
+
+        # Information desks
+        self.sync_geojson(lang, tourism_views.InformationDeskViewSet, 'information_desks.geojson')
+        for pk in tourism_models.InformationDeskType.objects.values_list('pk', flat=True):
+            name = 'information_desks-{}.geojson'.format(pk)
+            self.sync_geojson(lang, tourism_views.InformationDeskViewSet, name, type=pk)
+        for desk in tourism_models.InformationDesk.objects.all():
+            self.sync_media_file(lang, desk.thumbnail)
 
         self.zipfile.close()